requirements.txt
streamlit
torch
torchvision
captum
numpy
pandas
pillow
opencv-python-headless
matplotlib
scikit-image
scikit-learn
